Disable doclint in JDK 8 Javadoc
Doclint was introduced in JDK 8 and checks Javadoc comments to conform W3C HTML 4.01 HTML. In my opinion it's a little bit to strict per default and crashes documentations, that worked well. This is an easy way to disable it per default. Maybe this should be optional with a gradle property. If you want, I can insert that as well.